#6: Renamed actions.js file in auth module
