+++ app/src/store/modules/auth/index.js
@@ -1,5 +1,5 @@
 import mutations from './mutations.js';
-import actions from './actions.js.js';
+import actions from './actions.js';
 import getters from './getters.js';
 
 export default {
